requirements.txt
sqlalchemy==2.0.23
pymongo==4.6.0
pymysql==1.1.0
psycopg2-binary==2.9.9
sentence-transformers==3.3.1
pymilvus==2.5.3
python-dotenv==1.1.0
grpcio>=1.60.0 
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3